Significant changes in UsefulFunctions
Added comments on the analyze module.
Added comments on RecursiveLowPassFast
I fixed Upwards Spikes. Now pressing the button has the result. In the past, nothing happened at all.
Has removed recursive pass pass fast. I replaced it with the appropriate if in RecursiveLowPassFast.
Commented out the module SavingAndPlottingAfterRecursive Reorganized to make everything clear.
Suspicious code in the 730 line
The colorbutton has been activated in the lower left corner of the screen !!! Now you can really change colors.
Added a normal description in the plot Scatter plot !!! Prior to this, the logarithm of time was constructed along the x-axis at a logarithmic scale. A signature to the picture said that this is real time, just the scale of plotting the logarithmic. Caused to misunderstanding. Now it is immediately clear that we are building! Corrected the value of scatter plot-> timeline! Time is done in microseconds.
I fixed the Dwell schedule - it did not work now.
Put clear symbols on all graphs. Now it is clear in what and how boundaries change when plotting.